+++ pkg/agent/pyspy/pyspy.go
@@ -37,7 +37,7 @@ type PySpy struct {
 	pid int
 }
 
-func Start(pid int, _ spy.ProfileType, _ uint32, _ bool) (spy.Spy, error) {
+func Start(params spy.InitParams) (spy.Spy, error) {
 	dataBuf := make([]byte, bufferLength)
 	dataPtr := unsafe.Pointer(&dataBuf[0])
 
@@ -52,7 +52,7 @@ func Start(pid int, _ spy.ProfileType, _ uint32, _ bool) (spy.Spy, error) {
 	if Blocking {
 		blocking = 1
 	}
-	r := C.pyspy_init(C.int(pid), C.int(blocking), errorPtr, C.int(bufferLength))
+	r := C.pyspy_init(C.int(params.Pid), C.int(blocking), errorPtr, C.int(bufferLength))
 
 	if r < 0 {
 		return nil, errors.New(string(errorBuf[:-r]))
@@ -63,7 +63,7 @@ func Start(pid int, _ spy.ProfileType, _ uint32, _ bool) (spy.Spy, error) {
 		dataBuf:  dataBuf,
 		errorBuf: errorBuf,
 		errorPtr: errorPtr,
-		pid:      pid,
+		pid:      params.Pid,
 	}, nil
 }